src: add a stub for user-defined functions
The function doesn't do much now, but it can be successfully
instantiated from postgres. Bonus - it's in the same shared
library as the logical replication plugin, so we only need
to add the Postgres boilerplate once. If it ever proves to be
problematic, we can just move it to a separate shared lib.